Move DB API-client behind interface
