+++ Cargo.toml
@@ -14,3 +14,4 @@ quick-xml = { version = "0.23.0", features = ["serialize"] }
 chrono = { version = "0.4.19" }
 merge = { version = "0.1.0" }
 serde_json = { version = "1.0" }
+async-trait = { version = "0.1.60" }
